move more headers into namespace-specific subdir under ./include
