+++ include/concepts/board_state_calculator.hpp
