+++ include/concepts/board_state_coordinator.hpp
