+++ include/concepts/composite_concepts.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
-#include <concept_board_state_calculator.hpp>
+#include <concepts/board_state_calculator.hpp>
 #include <concepts/board_state_calculator_registry.hpp>
 #include <concepts/single_board_state_provider.hpp>
-#include <concept_space_info_provider.hpp>
+#include <concepts/space_info_provider.hpp>
 #include <concepts>
 
 template <typename T>
