+++ include/concepts/space_info_provider.hpp
