+++ tests/core/zobrist_calculator_concept_test.cpp
@@ -1,8 +1,8 @@
 #include <board_data_structs.hpp>
 #include <board_map_fixture.hpp>
 #include <builders.hpp>
-#include <concept_board_state_calculator.hpp>
-#include <concept_composite_concepts.hpp>
+#include <concepts/board_state_calculator.hpp>
+#include <concepts/composite_concepts.hpp>
 #include <concepts/single_board_state_provider.hpp>
 #include <gtest/gtest.h>
 #include <zobrist_calculator_for_concepts.hpp>
