+++ tests/core/zobrist_component_concept_test.cpp
@@ -1,7 +1,7 @@
 #include <array>
 #include <board_map_fixture.hpp>
 #include <builders.hpp>
-#include <concept_multi_board_state_provider.hpp>
+#include <concepts/multi_board_state_provider.hpp>
 #include <concepts/single_board_state_provider.hpp>
 #include <game_board_for_concepts.hpp>
 #include <gtest/gtest.h>
